Issue #4: Move ILLPortalURL to the configuration page
This patch introduces the use of the SNLP plugin as an object in the
SNLP backend construtor. It gets initializaed, as well as the
configuration for later use.

The configuration is then used to read the fresh *portal_url*
configuration, to properly render the URL in the create.inc file.

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>