This patch introduces the command SLNPPFLDatenAenderung to change incoming items attributes referenced by PFL number. This attributes could be - sigel_lending_library (SigelGB) -> to map the Koha::Patron - itemcallnumber (Signatur)
Readme is updated with new lending_library config section which defines where to look for the given SigelGB in Koha::Patrons

When pfl_number_prefix is set to @ the zfl server will omit it when sending a DatenAenderung requests. It will then send only the ill request id.
So when pfl_number_prefix is set to @ we check only for numerical values in param PFLNummer.
Else the whole prefix and number get evaluated.

The lending library will be displayed in illviews metadata